refactor: rename Geo to GeoMap temporarily due to casing issue with Github
