+++ giraffe/src/components/GeoMap/SVGIcon.ts
