+++ giraffe/src/components/GeoMap/bing-maps/Bing.ts
