+++ giraffe/src/components/GeoMap/bing-maps/BingLayerObject.ts
