+++ giraffe/src/components/GeoMap/bing-maps/BingMap.tsx
