+++ giraffe/src/components/GeoMap/dimensionCalculations.ts
