+++ giraffe/src/components/GeoMap/processing/GeoTable.ts
