+++ giraffe/src/components/GeoMap/processing/NativeGeoTable.ts
