+++ giraffe/src/components/GeoMap/processing/PivotedGeoTable.ts
