+++ giraffe/src/components/GeoMap/processing/mixins.ts
