+++ giraffe/src/components/GeoMap/processing/tableProcessing.ts
